refactor: refactor project
