+++ .dockerignore
@@ -1,1 +1,1 @@
-.env
+config.toml
