+++ .gitignore
@@ -4,6 +4,7 @@
 *.dll
 *.so
 *.dylib
+bin/*
 
 # Test binary, built with `go test -c`
 *.test
@@ -17,4 +18,5 @@
 # IDE
 .idea
 
-.env
+# Project
+configs/config.toml
