+++ deploy/docker-compose.yml
@@ -1,0 +1,9 @@
+version: "3"
+services:
+  capture-maps-api:
+    build: "../"
+    restart: unless-stopped
+    ports:
+      - 9002:8090
+    environment: 
+      GOOGLE_MAPS_API_KEY: "<GOOGLE_MAPS_API_KEY>"
