+++ examples/README.md
@@ -1,0 +1,1 @@
+[Postman Collection](https://www.getpostman.com/collections/ec8c981f5c2ad23a2531)
