--- src/earth/compute_test.go
@@ -1,7 +1,0 @@
-package earth
-
-import "testing"
-
-func TestSamples(t *testing.T) {
-
-}
