--- src/main_test.go
@@ -1,30 +1,0 @@
-package main
-
-import (
-	"net/http"
-	"net/http/httptest"
-	"strings"
-	"testing"
-)
-
-func TestHome(t *testing.T) {
-	req, err := http.NewRequest("GET", "/", nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	rr := httptest.NewRecorder()
-	handler := http.HandlerFunc(Home)
-	handler.ServeHTTP(rr, req)
-
-	if status := rr.Code; status != http.StatusMovedPermanently {
-		t.Errorf("handler returned wrong status code: got %v want %v",
-			status, http.StatusMovedPermanently)
-	}
-
-	expected := `<a href="http://zxcv32.com">Moved Permanently</a>.`
-	body := strings.TrimSpace(rr.Body.String())
-	if body != expected {
-		t.Errorf("handler returned unexpected body: got %v want %v",
-			rr.Body.String(), expected)
-	}
-}
