+++ test/earth_test/compute_test.go
@@ -1,0 +1,8 @@
+package earth_test
+
+import "testing"
+
+// Blackbox test
+func TestSamples(t *testing.T) {
+
+}
